split the models to a shared library.  started a console app that uses the shared library.
