+++ starterBash.sln
@@ -5,16 +5,25 @@ VisualStudioVersion = 15.0.28010.2041
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starterBash", "starterBash\starterBash.csproj", "{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bash-gen", "bash-gen\bash-gen.csproj", "{7BF9D3BA-AB50-44DD-B16C-753667C356F5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bashGeneratorSharedModels", "bashGeneratorSharedModels\bashGeneratorSharedModels.csproj", "{55458040-B7B9-4411-899A-282263A35D5C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|ARM = Debug|ARM
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|ARM = Release|ARM
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|Any CPU.Build.0 = Debug|x86
+		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|Any CPU.Deploy.0 = Debug|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|ARM.ActiveCfg = Debug|ARM
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|ARM.Build.0 = Debug|ARM
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|ARM.Deploy.0 = Debug|ARM
@@ -24,6 +33,7 @@ Global
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|x86.ActiveCfg = Debug|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|x86.Build.0 = Debug|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Debug|x86.Deploy.0 = Debug|x86
+		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|Any CPU.ActiveCfg = Release|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|ARM.ActiveCfg = Release|ARM
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|ARM.Build.0 = Release|ARM
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|ARM.Deploy.0 = Release|ARM
@@ -33,6 +43,38 @@ Global
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|x86.ActiveCfg = Release|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|x86.Build.0 = Release|x86
 		{1E4EA646-51F2-4215-9468-0E16BF4F7EB9}.Release|x86.Deploy.0 = Release|x86
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|ARM.Build.0 = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|x64.Build.0 = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Debug|x86.Build.0 = Debug|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|ARM.ActiveCfg = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|ARM.Build.0 = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|x64.ActiveCfg = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|x64.Build.0 = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|x86.ActiveCfg = Release|Any CPU
+		{7BF9D3BA-AB50-44DD-B16C-753667C356F5}.Release|x86.Build.0 = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|ARM.Build.0 = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|x64.Build.0 = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Debug|x86.Build.0 = Debug|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|ARM.ActiveCfg = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|ARM.Build.0 = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|x64.ActiveCfg = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|x64.Build.0 = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|x86.ActiveCfg = Release|Any CPU
+		{55458040-B7B9-4411-899A-282263A35D5C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
