+++ starterBash/MainPage.xaml
@@ -3,6 +3,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:starterBash"
+    xmlns:shared="using:sharedBashGen"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -25,7 +26,7 @@
             <Setter Property="TextAlignment" Value="Left" />
             <Setter Property="IsSpellCheckEnabled"  Value="False" />
         </Style>
-        <DataTemplate x:Key="ParameterItem" x:DataType="local:ParameterItem">
+        <DataTemplate x:Key="ParameterItem" x:DataType="shared:ParameterItem">
             <Grid Background="Gray" HorizontalAlignment="Stretch">
                 <Grid.RowDefinitions>
                     <RowDefinition Height="*" />
@@ -208,7 +209,8 @@
                 <TextBlock Text="Script File Name: " FontSize="20" Style="{StaticResource KvpTextBlock}" />
                 <TextBox FontSize="20" Margin="10,0,20,0" Width="450" VerticalAlignment="Center" Text="{x:Bind ScriptName, Mode=TwoWay}" Style="{StaticResource KvpTextBox}" GotFocus="Text_GotFocus" LostFocus="TextBox_LostFocus" />
                 <CheckBox Content="Echo Input" Margin="20,0,0,0" IsChecked="{x:Bind EchoInput, Mode=TwoWay}"/>
-                <CheckBox Content="Add Log Lines" Margin="20,0,0,0" IsChecked="{x:Bind CreateLogLines, Mode=TwoWay}"/>
+                <CheckBox Content="Create Log File" Margin="20,0,0,0" IsChecked="{x:Bind CreateLogFile, Mode=TwoWay}"/>
+                <CheckBox Content="Tee to Log file" Margin="20,0,0,0" IsChecked="{x:Bind TeeToLogFile, Mode=TwoWay}"/>
 
             </StackPanel>
             <ListBox Name="ListBox_Parameters" Grid.Row="1" ItemsSource="{x:Bind Parameters }" ItemTemplate="{StaticResource ParameterItem}" SelectionChanged="OnSelectionChanged" HorizontalContentAlignment="Stretch">
